+++ packages/minimongo/minimongo.js
@@ -694,21 +694,32 @@ LocalCollection._findInOrderedResults = function (query, doc) {
   throw Error("object missing from query");
 };
 
+// This binary search puts a value between any equal values, and the first
+// lesser value.
+LocalCollection._binarySearch = function (cmp, array, value) {
+  var first = 0, rangeLength = array.length;
+
+  while (rangeLength > 0) {
+    var halfRange = Math.floor(rangeLength/2);
+    if (cmp(value, array[first + halfRange]) >= 0) {
+      first += halfRange + 1;
+      rangeLength -= halfRange + 1;
+    } else {
+      rangeLength = halfRange;
+    }
+  }
+  return first;
+};
+
 LocalCollection._insertInSortedList = function (cmp, array, value) {
   if (array.length === 0) {
     array.push(value);
     return 0;
   }
 
-  for (var i = 0; i < array.length; i++) {
-    if (cmp(value, array[i]) < 0) {
-      array.splice(i, 0, value);
-      return i;
-    }
-  }
-
-  array.push(value);
-  return array.length - 1;
+  var idx = LocalCollection._binarySearch(cmp, array, value);
+  array.splice(idx, 0, value);
+  return idx;
 };
 
 // To track what documents are affected by a piece of code, call saveOriginals()